Ajout de la fonction recolte des villageois

+++ src/modele/unite/structure/Recoltable.java
@@ -1,5 +1,5 @@
 package modele.unite.structure;
 
 public interface Recoltable {
-    int enlever(int qte);
+    double enlever(double qte);
 }
